Allow for middleware to pre-process values
Middleware allows for all parameters to be processed
and replaced before being passed on to sprintf. An example
of where this is useful is if most or all parameters need
to have escapeshellarg() ran on them. Instead of running
escapeshellarg() manually on each parameter, middleware can
be passed in once and Sprintf::sprintf() can do the work
of calling it on each parameter.