Switch regex to use nowdoc
In certain environments (such as http://sandbox.onlinephpfunctions.com/),
`preg_match` complains if the pattern string contains a newline after
the pattern modifiers.  Using a nowdoc suppresses the end-of-string
newline.

```
Warning:  preg_match(): Unknown modifier '
'
```

No tests are being included with this fix because I could not reproduce
the error in the Docker or Travis CI environments.

Resolves #61